AMD-microservice/Items-service/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-jose
python-multipart